Refactor index so app can be exported
I wanted to start a console with our app loaded for easier debugging of
stuffs.

To do this I need app to be available for export without starting the
server.

This commit moves the app definition into its own file which can be
required separately
