+++ src/index.js
@@ -1,21 +1,9 @@
-import cors from 'express'
-import express from 'express'
-import sqlite from 'sqlite3'
-import ejs from 'ejs'
+import app from './app.js'
 
-const db = new sqlite.Database('../db.sqlite')
-
-const app = express()
 const port = 4004
-
-app.set('view engine', 'ejs')
-
-app.get('/', function(request, response) {
-  response.render('index')
-})
 
 app.listen(port, function() {
   console.log(`Listening on port ${port}`)
 })
 
-db.close()
+app.db.close()
